// src/app/models/priest.model.ts
export class Priest {
    id?: any;
    firstname?: string;
    secondname?: string;
    lastname?: string;
    secondlastname?: string;
}

export class Bpriest {
    id_priest: string
    firstname_priest: string
    secondname_priest: string
    lastname_priest: string
    second_lastname_priest: string
    credentials: string
    parish_origin: string
}